ip_maven/src/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
kombu==5.4.1
numpy==2.1.1
packaging==24.1
pandas==2.2.2
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
watchdog==5.0.2
wcwidth==0.2.13
